Move old disjoint tests
And address other comments
