+++ tests/test_util.py
@@ -28,31 +28,85 @@ import logging
 import msprime
 import numpy as np
 import pytest
+import tsinfer
 import tskit
 
 import tsdate
 
 
 class TestSplitDisjointNodes:
+    @staticmethod
+    def has_disjoint_nodes(ts):
+        """
+        Brute force check for disjoint nodes, by pulling out edge intervals for
+        each node; taking the union of intervals; checking that a single
+        interval remains
+        """
+
+        def merge_intervals(intervals):
+            intervals = sorted(intervals, key=lambda x: x[0])
+            result = []
+            (start_candidate, stop_candidate) = intervals[0]
+            for start, stop in intervals[1:]:
+                if start <= stop_candidate:
+                    stop_candidate = max(stop, stop_candidate)
+                else:
+                    result.append((start_candidate, stop_candidate))
+                    (start_candidate, stop_candidate) = (start, stop)
+            result.append((start_candidate, stop_candidate))
+            return result
+
+        intervals_by_node = {i: [] for i in range(ts.num_nodes)}
+        for e in ts.edges():
+            intervals_by_node[e.parent].append([e.left, e.right])
+            intervals_by_node[e.child].append([e.left, e.right])
+
+        for n in range(ts.num_nodes):
+            intr = merge_intervals(intervals_by_node[n])
+            if len(intr) != 1:
+                return True
+
+        return False
+
+    @staticmethod
+    def childset_changes_with_root(ts):
+        """
+        If root nodes are split whenever their children change, the next root
+        should have the same child set if it has the same ID
+        """
+        last_childset = frozenset()
+        last_root = tskit.NULL
+        for t in ts.trees():
+            if t.num_edges == 0:
+                last_childset = frozenset()
+                last_root = tskit.NULL
+            else:
+                if t.num_roots > 1:
+                    return False
+                childset = frozenset(list(t.children(t.root)))
+                if t.root == last_root and childset != last_childset:
+                    return False
+                last_childset = childset
+                last_root = t.root
+        return True
+
     def test_nosplit(self):
         ts = tskit.Tree.generate_comb(5).tree_sequence
         split_ts = tsdate.util.split_disjoint_nodes(ts)
-        assert split_ts.num_nodes == ts.num_nodes
-        assert split_ts.num_edges == ts.num_edges
-        assert split_ts.num_trees == ts.num_trees
-        for node in split_ts.nodes():
-            assert node.flags & tsdate.NODE_SPLIT_BY_PREPROCESS == 0
+        assert ts.equals(split_ts, ignore_provenance=True)
         prov = json.loads(split_ts.provenance(-1).record)
         assert prov["software"]["name"] == "tsdate"
         assert prov["parameters"]["command"] == "split_disjoint_nodes"
 
-    def test_simple(self):
+    def test_simple(self, caplog):
         tables = tskit.Tree.generate_comb(5).tree_sequence.dump_tables()
         tables.delete_intervals([[0.2, 0.8]])
         tables.nodes.metadata_schema = tskit.MetadataSchema.permissive_json()
         ts = tables.tree_sequence()
         num_internal_nodes = ts.num_nodes - ts.num_samples
-        split_ts = tsdate.util.split_disjoint_nodes(ts)
+        with caplog.at_level(logging.WARNING):
+            split_ts = tsdate.util.split_disjoint_nodes(ts)
+            assert caplog.text == ""
         num_new_internal_nodes = split_ts.num_nodes - split_ts.num_samples
         assert split_ts.num_nodes > ts.num_nodes
         # all internal nodes should be split
@@ -130,6 +184,23 @@ class TestSplitDisjointNodes:
         assert split_ts.num_provenances == ts.num_provenances
         split_ts = tsdate.util.split_disjoint_nodes(ts, record_provenance=True)
         assert split_ts.num_provenances == ts.num_provenances + 1
+
+    def test_inferred(self):
+        ts = msprime.sim_ancestry(
+            10,
+            population_size=1e4,
+            recombination_rate=1e-8,
+            sequence_length=1e6,
+            random_seed=1,
+        )
+        ts = msprime.sim_mutations(ts, rate=1e-8, random_seed=1)
+        sample_data = tsinfer.SampleData.from_tree_sequence(ts)
+        inferred_ts = tsinfer.infer(sample_data).simplify()
+        split_ts = tsdate.util.split_disjoint_nodes(inferred_ts)
+        assert self.has_disjoint_nodes(inferred_ts)
+        assert not self.has_disjoint_nodes(split_ts)
+        assert split_ts.num_edges == inferred_ts.num_edges
+        assert split_ts.num_nodes > inferred_ts.num_nodes
 
 
 class TestPreprocessTs:
